\name{ag.contourplot.ds}
\alias{ag.contourplot.ds}
\title{Computes a grid-density object required to generate a contour plot.}
\usage{
  ag.contourplot.ds(xvect, yvect)
}
\arguments{
  \item{xvect}{a numerical vector}

  \item{yvect}{a numerical vector}
}
\value{
  a heatmap plot.
}
\description{
  Computes a grid-density object required to generate a
  contour plot.
}
\author{
  Burton, P.
}

