\name{ag.histogram.ds}
\alias{ag.histogram.ds}
\title{Computes a histogram of the given data values without plotting.}
\usage{
  ag.histogram.ds(a)
}
\arguments{
  \item{a}{a vector of values for which the histogram is
  desired.}
}
\description{
  Computes a histogram of the given data values without
  plotting.
}
\author{
  Gaye A.
}

