\name{ag.MinMax.ds}
\alias{ag.MinMax.ds}
\title{Generates a matrix with a minimum and maximum value, for two numerical vector}
\usage{
  ag.MinMax.ds(xvect, yvect)
}
\arguments{
  \item{xvect}{a numerical vector}

  \item{yvect}{a numerical vector}
}
\value{
  a matrix
}
\description{
  This function used by another function to plot a heatmap
  graph.
}
\author{
  Isaeva, J.
}

