\name{ag.odds.ratio.ds}
\alias{ag.odds.ratio.ds}
\title{Calculates odds ratio and its 95 percent confidence interval}
\usage{
  ag.odds.ratio.ds(xvect, yvect, alpha = 0.05)
}
\arguments{
  \item{xvect}{a factor vector}

  \item{yvect}{a factor vector}

  \item{alpha}{level of statistical significance}
}
\description{
  Calculates odds ratio and its 95 percent confidence
  interval
}
\author{
  Isaeva J.
}

