\name{ag.product.ds}
\alias{ag.product.ds}
\title{Computes a product}
\usage{
  ag.product.ds(a = 1, b = 1, c = 1, d = 1, e = 1)
}
\arguments{
  \item{a}{a numerical}

  \item{b}{a numerical}

  \item{c}{a numerical}

  \item{d}{a numerical}

  \item{e}{a numerical}
}
\value{
  a numerical
}
\description{
  Computes a product
}
\author{
  Burton, P.
}

