\name{ag.quantile.mean.ds}
\alias{ag.quantile.mean.ds}
\title{Provides quantiles and mean information without maximum and minimum}
\usage{
  ag.quantile.mean.ds(a)
}
\arguments{
  \item{a}{a numerical vector}
}
\description{
  Provides quantiles and mean information without maximum
  and minimum
}
\author{
  Burton P.
}

