\name{ag.quantilemean.ds}
\alias{ag.quantilemean.ds}
\title{Provides quantiles and mean information without maximum and minimum}
\usage{
  ag.quantilemean.ds(a)
}
\arguments{
  \item{a}{a numerical vector}
}
\description{
  Provides quantiles and mean information without maximum
  and minimum
}
\author{
  Burton, P.
}

